now csv's table is a separated component
